destination-s3-glue: Fix decimal type syntax
The definition of the decimal type as `decimal(38,38)` was incorrect as it caused the
query engine to interpret that as meaning that all 38 digits had to be to the right of
the decimal place, so any whole number values would overflow. Setting it as
`decimal(38)` allows for the full 38 digits, but with flexible scale.