Fix numeric types in S3 Glue
This updates the integer type used to be `bigint` unilaterally. This expands the range
of values so that they are not artificially constrained, since JSON integers don't have
any similar guarantees.

This implements the logic for `getNumericType` in the metastore format config so that
the type definition as it needs to be represented in the Glue catalog is returned,
removing the need for conditional logic within the schema transformation block.