airbyte-to-flow: set schema type when removing enums
When normalizing collection schemas, any `enum` keywrods are removed.
This creates problems when the schema doesn't otherwise specify a `type`
constraint, because then the schema is left with no constraints at all.
This causes fields to not be included in default materializations.
This commit changes the behavior to ensure that there is a `type`
whenever an `enum` is removed. The `type` is determined based on the
types of the values in the `enum`.