Sort property names in reverse order
Mixpanel's native properties are all lowercase. If there is a custom string property "Time" that conflicts with Mixpanel's numeric property "time", we should rename the custom one from "Time" to "_Time". Current behavior results in an exception trying to parse a string value when it expects a numeric one, see example line 156 of export.py.