[airbyte-ci] fix test regression following secret mounting strategy change
<!--
Thanks for your contribution! 
Before you submit the pull request, 
I'd like to kindly remind you to take a moment and read through our guidelines
to ensure that your contribution aligns with the type of contributions our project accepts.
All the information you need can be found here:
   https://docs.airbyte.com/contributing-to-airbyte/

We truly appreciate your interest in contributing to Airbyte,
and we're excited to see what you have to offer! 

If you have any questions or need any assistance, feel free to reach out in #contributions Slack channel.
-->

## What
A test is not passing as it always assumes the secrets are scrubbed by Dagger.
It's not the case since previous version: we now locally mount secrets as files, not secret.
The scrubbing does not run locally, the test won't pass if it's context is not set to CI.

## How
Set the context to CI on this specific test.