airbyte-to-flow: set resource_path in discover output
The `resource_path` field was recently added to the discovered response, to
facilitate better behavior when merging existing and discovered bindings.  This
commit updates airbyte-to-flow to start returning the `resource_path` along
with the discovered response.

Additionally, it updates how the `resource_path` is determined, to account for
the optional `namepsace` property of an airbyte stream.  This is intended to
ensure correct merge behavior in the case that multiple streams have the same
`name`.