remove debezium-bundles from deps.toml (#33267)
in deps.toml, we define a debezium-bundle which include all the debezium dependencies for all databases. That's being used in several subprojects, including CDK, which means CDK also depends on specific databases (mysql, postgres, mongo, mssql). As I want to remove that dependency in the future, this is preparing this by doing a no-op except for refactoring and clarifying dependencies. Note that mongo used to import debezium-bundle, and rather than risking a regression, I'm still importing all corresponding debezium libraries there.
The CDK still has a bunch of uses of DB-specific debezium libraries, so I left those there too.

I haven't bumped the CDK version nor the mongo connector version because this is a no-op. I will do that if the CI yells at me or if anyone feels strongly about it