improve airbyte-cdk (#35720)
various small changes related to tests and mssql:
the jdbc string for MSSQL was wrong. 
The LoggingInterceptor was printing faulty log lines
The SSHBastionContainer should be a container like any other, relying upon the ContainerFactory.
As a result, the ContainerFactory should be able to return any type of container, not just JdbcDatabaseContainer.
In addition, we shouldn't be passing strings to the ContainerFactory as container modifiers, but a named lambda. No more introspection needed! (the old string-based is kept for backward compatibility)

The CdcSourceTest doesn't check DB invariants before starting a source. We need a test function that can be overriden by the implementations of CdcSourceTest