add a timeout to junit test (#35767)
Today there's no clear timeout for tests. The only one that exists is at the github action level, and it has to be conservative at 1h.
This change introduces a timeout per test method. The default value is set at 5 minutes per method, and can be overriden on a per method basis via the use of the junit @Timeout annotation.

We're also fixing a regression in the test ContainerFactory that was sharing less container than was possible. Finally, we're introducing a MdcScope.doNothing to allow using MdcScope in try blocks without triggering a compiler warning