Fix tab delimiter configuration in CSV file type
The tab delimiter is only allowed to be a single character long. In the web UI entering
a literal tab character (`\t`) yields an escaped tab character (`\\t`) in the stored
configuration. This ensures that the escaped tab character is passed through to the call
for `csv.register_dialect` as the proper literal tab `\t`.