Fixing an error in code that checks custom GAQL queries
The issue here is with check for incremental queries. Some tables have a
limited lookback time frame to be valid. For instance, click_view should
not be asked for data older than 90 days.

Hardcoded 1980-01-01 is much older than 90 days. As the fix for the
check procedure I've added a much more recent time frame: last month.

The period is calculated based on today's date during the check.

Also it would be nice to see the exact query being executed during the
check in the logs.