allow BaseTypingDedupingTest to be interrupted (#37854)
When timing out, the LoggingInterceptor attempts to interrupt the test-executing thread. Unfortunately, the `CompletableFuture.join()` call is non-interruptible, which means we still had tests that were hanging forever. This changes it with `CompletableFuture.get()`, which is interruptible (even though the difference in behavior is not clearly documented. One has to notice that `get()` throws an `InterruptedException` while `join()` doesn't).