CDK: make postgres temp_file_limit error a transient error (#38734)
Fixes https://github.com/airbytehq/airbyte/issues/27090

Postgres throws a temp_file_limit error is translated to a system error for Airbyte. I believe this is true for both Postgres source and destination connectors.

This change makes this error transient so that it won't trigger sentry. It also fixes a bug that should use a lower-case version of a transient error message.