cdk: fix when "anyOf" entry doesn't get extracted into "type" values
Previously the only case that was handled when "anyOf" entry
was 2 items with one of them being "null". In any other "anyOf"
case no cleanup would happen from "anyOf" into simple "type" entries,
and the rest of the code expected a "type" entry to be available

Now only do "type" entry manipulation if it exists, which
would be the case of "anyOf" not being present anymore.