docs: increase the max partial failure attempt to 20. (#39322)
We are increasing the max partial failure attempt to 20 to accommodate the new CDC load algorithmn.

This may result in longer jobs. However all jobs that are longer will be moving data, so we'll also increase the chance a job succeeds in the event of transient errors.

Change was made in airbyte-platform. This is the docs update.