feat: pushes airbyte to Glide GCP Artifact registry
The readme contains instructions for how to use this with Airbyte and how to push new versions.
There were also some fixes to our airbyte config to make the API path and host optional and they default to prod.
This is probably getting close to closing glideapps/glide#27833 once it is reviewed