airbyte-ci: do not push latest tag for release candidates (#44551)
## What
Closes https://github.com/airbytehq/airbyte-internal-issues/issues/9403
When a connector is a release candidate it will have the `releases.isReleaseCandidate` flag set to `true` in its `metadata.yaml`.
We don't want to publish the `latest` tag when the connector is a RC.

## How
Tweak the publish pipeline to conditionnaly publish the latest tag according to the metadata content.