Reduce slice size to 7 days at minimum
Instead of 90-day slices, reduce it to a minimum of 7 days so that large data set still sync
without running into a lot of errors like 'Response ended prematurely'.

Also fixed a bug in how the range is adjusted. It was previously adjusting the range after
each record is processed which is way too frequent. It is now adjusted after each slice is
processed.