source-oracle: use CDK 0.34 and latest airbyte-ci (#43)
## What

With https://github.com/airbytehq/airbyte/pull/45360 the
airbyte-enterprise repo no longer needs to maintain docs files which
only existed to satisfy the QA checks. Accordingly, the `sl` value for
source-oracle has been set to 0 and the docs URL updated in
metadata.yaml.

This PR replaces the docs with a symlink. The source of truth for
enterprise connector documentation is now unambiguously in
airbytehq/airbyte. The symlink, however, confuses airbyte-ci, hence
https://github.com/airbytehq/airbyte/pull/45393 which needs to merge for
this PR to pass CI.

This PR also tracks the latest CDK changes.

## How
n/a

## Review guide
n/a

## User Impact
None

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [ ] YES 💚
- [x] NO ❌