destination-s3: fix issue with columnless streams (#45402)
Added support for handling empty data fields in Avro records and introduced a new test case for streams with no columns.

We've agreed there's a bug in a source (uscensus) that returns a single columnless stream. There's also a bug in the platform that allows a customer to create a connection using that columnless empty stream.
While it would be ideal for the destination connector to send an (yet non existent) `UpstreamError` back to the platform, we have some configurations of S3 (CSV, Json) that just allow teh columnless records to be persisted. We're just bringing the parquet/avro to the same permissiveness