source-oracle: add SSL and NNE integration tests (#57)
## What
This PR adds the missing test coverage for connections involving SSL or
NNE.

This PR also fixes the list of supported NNE encryption algorithms,
adding all supported variants, minus RC4 which is deprecated. For some
reason, regular DES is still supported (but not encouraged, but still),
which is nuts, but whatever.

This PR does NOT add support for SSL client authentication. This is
future work, to be done on an as-needed basis.

## How
Lots of testcontainer shenanigans to get the Oracle testcontainer to
self-sign a certificate and use that for SSL connections.

## Review guide
There's two commits, one for SSL, another for NNE

## User Impact
Nothing significant

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [x] YES 💚
- [ ] NO ❌