Update Hubspot Forms to Include all Form Types
The forms API defaults to only retrieving "hubspot" type forms. However, there is captured, flow and blog_comment type forms that can also be taken, these need to be explicitely added as formTypes to be collected. This is necessary for form submissions as well, as form submissions can come from all these types.