feat: make Service Now available in Cloud and OSS (#85)
## What

This PR renames `source-blackbird` to `servicenow`! 

It needs #84 to be merged before this can get in, and it depends on the
Compose license checks.

Gotchas I've learned doing this:
- TEST_CREDS should not be renamed as it refers to a GSM secret key.
- Don't forget to make it available in `metadata.yaml`
- Don't forget `ab_internal.isEnterprise` in `metadata.yaml`
- Docs problem: I don't know how validations will behave, BUT we should
have the doc files. They won't work from this repository, we will have
to either put the docs in the main repo, or make docusaurus fetch
additional doc files from github or relative directory.


## Next Steps
- [ ] Does this build and pass tests?
- [ ] Figure out documentation
- [ ] Test with Compose
- [ ] Ship the rest