[source-volcano] implementing a simple class for remote Hana database and implement tests for metadata querier to show its usage (#93)
This patch added a class that represents a remote SAP HANA database and
added tests for the metadata querier to show how to use it.

Every time a test is kicked off, a schema with UUID will be created as
its suffix. This way allows multiple integration tests to run without
interrupting each other. It is similar to implementing our CAT test
python hook.py for open-source connectors such as MySQL and Postgres.