chore: rename firetruck to oracle-enterprise (#97)
## What

This PR renames firetruck to Oracle Enterprise in everything that is
customer facing, but keeps the actual code the same. _In theory_ this
should be enough to make Oracle available with an Enterprise license.


## Is this safe to merge? 

When we merge this, we will have a new "enterprise" connector in the
registry. It won't show up for all users without enterprise license
feature flag, that @matteogp has already verified.

So all in all this is safe to merge and start experimenting with. As
long as I did not screw up the build, that's all it should take.

---------

Co-authored-by: Yue Li <yue.li@airbyte.io>