(fix: Source-Workday): Add date and datetime datatypes in schema_helper (#109)
## What
<!--
* Describe what the change is solving. Link all GitHub issues related to
this change.
-->
Solves: https://github.com/airbytehq/oncall/issues/7533

## How
<!--
* Describe how code changes achieve the solution.
-->
User gets `No xsd:date element found` when attempting to setup a RAAS
stream. We are assuming this is due to the TYPE_MAP not existing in
schema_helper.py - not 100% the logs provided were very limited but
think this is a good starting point.

## Review guide
<!--
1. `x.py`
2. `y.py`
-->

## User Impact
<!--
* What is the end result perceived by the user?
* If there are negative side effects, please list them.
-->

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [ ] YES 💚
- [ ] NO ❌