✨ Destination Cobra: support delete (#125)
## What
Support delete

## How
Enabling the choice in the spec. The rest should trickle down to the
code and work just fine.

Note that we still assemble the records even though there is just the ID
to send and there is not assembling to be done. The consequence is one
HTTP request per SF objects which seems minimal for now so we can leave
it.

## Review guide
<!--
1. `x.py`
2. `y.py`
-->

## User Impact
Users can select `DELETE` in the dropdown.

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [X] YES 💚
- [ ] NO ❌

Yes although the previously created config with `DELETE` might fail