Enabling support for other SASL mechanisms when SASL_PLAINTEXT protocol is chosen
Currently, the backend supports multiple SASL mechanisms with the SASL_SSL protocol. However, on the frontend, when SASL_PLAINTEXT is selected, there is no dropdown to choose the SASL mechanism—instead, it defaults to PLAIN behind the scenes, which is inconsistent. Since SASL_PLAINTEXT and SASL_SSL are nearly identical, differing only in their use of SSL for encryption during transit, this behavior doesn't make much sense. Adding new values to the enum has no negative impact and all tests continue to pass.