🐛 Destination Cobra: fix to have only one batch per job (#128)
## What
Sync that triggered this issue:
https://cloud.airbyte.com/workspaces/3dba3287-64bc-4a22-9692-0a4660604719/destination/1965b5eb-c003-49b3-8381-37c6d72adf83

The current implementation was done under the assumption that we could
post multiple batches for the same job. It turns out that doing so
generate the error `[{'errorCode': 'BULK_API_ERROR', 'message': "Found
multiple contents for job: <750P100000NpJI3>, please 'Close' / 'Abort' /
'Delete' the current Job then create a new Job and make sure you only do
'PUT' once on a given Job."}]`.

On the first attempted fix, we tries to scheduled multiple jobs at the
same time but we ended up with error `Error:UNABLE_TO_LOCK_ROW:unable to
obtain exclusive access to this record:--`. Hence we had do to a bit of
refactor to only have one job per stream at the same time. Note that if
multiple streams update the same sf object, we might still have the same
error. We have a warning log telling the user what to do in that case.

We also added a commit after the user complained of the many jobs being
created. Basically, the limit in terms of number of records only applies
on the Bulk API V1 and we're using V2. Hence the limit should be the
number of bytes. We implemented this as part of this PR.


## How
Have only one batch per job.


https://airbytehq-team.slack.com/archives/C08BGNKAL21/p1743774016167529?thread_ts=1743532473.710939&cid=C08BGNKAL21

Sorry @aaronsteers, this PR is a bit of a mess but it changes the core
flow of the way we create jobs so it is quite big...

## Review guide
<!--
1. `x.py`
2. `y.py`
-->

## User Impact
<!--
* What is the end result perceived by the user?
* If there are negative side effects, please list them.
-->

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [X] YES 💚
- [ ] NO ❌

---------

Co-authored-by: Natik Gadzhi <natik@respawn.io>