Update manifest.yaml for the GET_FBA_FULFILLMENT_REMOVAL_SHIPMENT_DETAIL_DATA lookback_window parameter
This particular report is missing data often due to the nature Amazon generates this report. dataEndTime can't be a proper cursor without the lookback_window parameter set to at least 31 days. What Amazon does is it "ends" the report at one date but do the actual shipments sometimes few weeks later. These shipments will be missed from the sync as their dataEndTime will be long time before the cursor especially with daily syncs enabled. To have this shipments syncs we need to look back for that shipments. They other fix to consider would be to maybe change the actual cursor to a shipment-date field maybe, but there's no guarantee they are all synced into the system at the same time at different fulfillment centers. Of course my proposal has to be used with the proper dedup in place, but for that to happen user also need to update the primary key to add one more field, the "disposition" as there might be two identical SKUs within the same removal order and in the same shipment but with two different dispositions.