[SAP HANA] Enable trigger log retention management (#158)
## What
This PR allow use to be able to delete trigger logs that has been
committed to the destination so we can free up space from the customer's
database

## How
The incremental sync runs based on SapHanaJdbcCursorIncrementalPartition
and that include cdc and user defined cursor. In partition reader, if we
see it's SapHanaJdbcCursorIncrementalPartition and it's cdc sync (for
sap hana it's trigger based CDC for sure atm), we will do run a delete
query in trigger table to do the clean up. The delete query will delete
all log record before the cursor checkpoint that's in the stream state.
Then it will proceed with normal incremental sync.

The delete query will be executed again upon the previous sync failure,
but that should be ok because the checkpoint shouldn't updated after the
failure. It will basically delete deleted records again which means
doing nothing.

## Review guide
<!--
1. `x.py`
2. `y.py`
-->

## User Impact
<!--
* What is the end result perceived by the user?
* If there are negative side effects, please list them.
-->

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [ ] YES 💚
- [ ] NO ❌