chore: correct Gradle caching behaviour for JVM connector builds. (#60844)
The current configuration writes on all branches. This can lead to confusing behaviour since we don't kill previous test runs, so the scenario where an earlier build writes the cache that a later build uses is possible. We've also run into some hard-to-debug caching weirdness we believe is related, so let's simplify the configuration to only write on master and read on PR builds.