fix: only build and push connectors if versions have changed. (#61584)
What
A bug currently exists in the publishing flow for JVM connectors - we always build and publish connectors on changes.

This is fine if the connector's version actually changed. Some destinations use requireVersionIncrementsInPullRequests to false to allow for rapid iteration without publishing versions.

Because of this, we want to add logic here to only publish images if the image does not exists.

How
Curl dockerhub to see if the image exists. If it does, do not publish.

I also added a --publish flag and defaulted the script to dry-run for better testing. This was previously requested, and I though I'd do so since I was here.