Updated node selector syntax to reflect that scheduling nodes are not automatically inherited as I expected. Thanks Peter.
Peter noticed it actually was mandatory to specify scheduling, discover, and spec nodeSelectors. Otherwise, they'll run in the static node group, which we don't want. So inheritance is only implied, but not actual. The specific node must be set for each. I'm also adding an entry for spec BUT this is currently missing from the chart. It needs to be added, and I'm documenting it here on the assumption that it will be added. However, technically, as of today, it doesn't work.