CDK version use the SemVer format (#63740)
This PR is making the CDK to use SemVer.

It requires the contributor to manually set the CDK version allowing to define patch/minor/major version.

This is a first step into improving the way we manage the CDK versioning.

The changes are the following ones:

java-bulk-cdk-publish.yml: remove the autogeneration of the version. Add a check step to ensure that we don't override the latest version
java-cdk-tests.yml: add a step in the CDK CI to ensure that the we don't override existing version.
README.md: Update the doc
build.gradle: remove the task which generate the build number; add a task to check the version doesn't conflict with the latest version.


Co-authored-by: Davin Chia <davinchia@gmail.com>