Update how we check the cdk version (#64567)
We suddenly start seeing errors in the checkCdkVersion task.
The error is related to the use of a deprecated java library.
There is no obvious change in this repo but there might be a change of the JVM version in the CI executor.

This is an example of a fail build related to that: https://github.com/airbytehq/airbyte/actions/runs/16835803871/job/47694907595