Update connector migration guides to use a reusable block of text instead of a link.
The UX in the app requires the user to click a link to get to the migration page, so we judged it a bad experience to then immediately ask them to go to another page again. This way still allows for a reusable set of text that can be easily updated, but the end user UX is less demanding.