Jwt Token Implementation
API Key implementation for DoorDash is nearing its end of life service so we need to switch over to jwt token based.
Note:  We could not make any changes in airbyte to generate a jwt token due to cryptography library incompatibly.  Instead we are calling service in GC to generate a jwt and return to aribyte.