Destination Snowflake: do not upcase column name in DROP COLUMN (#67582)
## What
for https://github.com/airbytehq/oncall/issues/9621. Branched from
https://github.com/airbytehq/airbyte/pull/67575. Stop upcasing the
column name before dropping it.

This only really comes into play if either (a) a schema evolution is
interrupted, or (b) the user manually adds a column with lowercase name
to their table. Most schema evolutions are working fine, in spite of
this bug.

## How
The bug is that when we `describe table`, we called
`toSnowflakeCompatibleName()` on the column names. This upcases the
column names. Then when we go to drop the column, the column obviously
doesn't exist, because we modified it incorrectly.

This PR switches to the method which does the quote-escaping, but
doesn't upcase the name.

Tested manually: 
```sql
alter table edgao_test.variant_test add column "x" object;
alter table edgao_test.variant_test add column "y" object;
```
Then run this thing to simulate a sync (the sync will create a table
with column `X`). Without this change, it fails; with this change, it
correctly drops the `x` and `y` columns, and leaves the `X` column
untouched.
```kotlin
    @Test
    fun arst() {
        val stream =
            DestinationStream(
                "edgao_test",
                "variant_test",
                importType = Append,
                schema =
                    ObjectType(
                        properties =
                            linkedMapOf(
                                "x" to FieldType(ObjectType(properties = linkedMapOf()), nullable = true)
                            )
                    ),
                generationId = 42,
                minimumGenerationId = 0,
                syncId = 42,
                namespaceMapper = namespaceMapperForMedium()
            )
        runSync(
            updatedConfig,
            stream,
            listOf(
                InputRecord(
                    stream,
                    """
                    {
                      "x": {"foo": "bar"}
                    }""".trimIndent(),
                    emittedAtMs = 1000,
                    checkpointId = checkpointKeyForMedium()?.checkpointId
                )
            )
        )
    }
```

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [x] YES 💚
- [ ] NO ❌