docs: clean up IP allow list URLs and security info re: data planes (#67074)
## What
<!--
* Describe what the change is solving. Link all GitHub issues related to
this change.
-->

This is minor docs cleanup after we released the Iowa region on GCP.
There's not much to update, but it did prompt me to take a look at a few
things, update some URLs, and reframe the discussion of how Airbyte is
deployed to Cloud.

Companion change to this UI update, but there is no dependency between
them. https://github.com/airbytehq/airbyte-platform-internal/pull/17923

## How
<!--
* Describe how code changes achieve the solution.
-->

## Review guide
<!--
1. `x.py`
2. `y.py`
-->

## User Impact
<!--
* What is the end result perceived by the user?
* If there are negative side effects, please list them. 
-->

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [x] YES 💚
- [ ] NO ❌

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>