Bump cdk for SF and release. (#67599)
## What
Brings SF to the latest CDK

## Notes
We no longer set max concurrent stream to 50 which is a good thing as
that could have lead to massive disk usage. We will now evict if we use
over 1750mb (350mb * 5) of disk, which is much more reasonable.

## User Impact
We won't immediately evict aggregates if over some static number of
interleaved-ness. We secondarily check the size of all active aggregates
and only evict if over the configured size.