🐙 source-7shifts: run up-to-date pipeline [2025-10-14] (#65634)
# Update source-7shifts

This PR was autogenerated by running `airbyte-ci connectors
--name=source-7shifts up_to_date --pull`

We've set the `auto-merge` label on it, so it will be automatically
merged if the CI pipelines pass.
If you don't want to merge it automatically, please remove the
`auto-merge` label.
Please reach out to the Airbyte Connector Tooling team if you have any
questions or concerns.


## Operations

- Upgrade the base image to the latest version in metadata.yaml:
Successful

- PATCH bump source-7shifts version to 0.0.29: Successful

- Build source-7shifts docker image for platform(s) linux/amd64,
linux/arm64: Successful

- Get dependency updates: Successful

- Create or update pull request on Airbyte repository: Successful

- Add changelog entry: Successful




## Dependency updates

We use [`syft`](https://github.com/anchore/syft) to generate a SBOM for
the latest connector version and the one from the PR.
It allows us to spot the dependencies that have been updated at all
levels and for all types of dependencies (system, python, java etc.).
Here are the dependencies that have been updated compared to
`airbyte/source-7shifts:latest`.
Keep in mind that `:latest` does not always match the connector code on
the main branch.
It is the latest released connector image when the head commit of this
branch was created.

| Type | Name | State | Previous Version | New Version |
|------|------|-------|-------------|------------------|
| python | airbyte-cdk | updated | 6.60.0 | **7.3.4** |
| python | anyio | updated | 4.9.0 | **4.11.0** |
| python | cattrs | updated | 24.1.2 | **25.2.0** |
| python | importlib_metadata | updated | 8.0.0 | **8.7.0** |
| python | joblib | updated | 1.5.1 | **1.5.2** |
| python | jsonschema | updated | 4.17.3 | **4.25.1** |
| python | pandas | updated | 2.2.2 | **2.2.3** |
| python | whenever | updated | 0.6.16 | **0.8.8** |
| python | zipp | updated | 3.19.2 | **3.23.0** |
| python | jsonschema-specifications | added | not present |
**2025.4.1** |
| python | referencing | added | not present | **0.36.2** |
| python | rpds-py | added | not present | **0.27.1** |
| python | filelock | removed | 3.18.0 | **not present** |
| python | pyrsistent | removed | 0.20.0 | **not present** |

Co-authored-by: octavia-bot-hoard[bot] <230633153+octavia-bot-hoard[bot]@users.noreply.github.com>