ci: run cdk compatibility test only on certified connectors (#67604)
## What
The existing `CDK Connector Compatibility Test` workflow currently tests
all connectors using the bulk CDK. However, we only want to test
certified connectors. This PR modifies the workflow to run tests
exclusively on certified connectors using the CDK.

## How
Instead of getting the list of connectors to test from the diff
generated by running a CDK upgrade task, we will get the list of
certified connectors by reusing the script used in the auto bump
workflow to get this information.

## How did I test this?
Tested it in [this](https://github.com/airbytehq/airbyte/pull/67605) PR.
More specifically, here's the PR check run
https://github.com/airbytehq/airbyte/actions/runs/18410594109/job/52461620803?pr=67605

Note: it looks like the connectors we are testing didn't change. I'd
guess it is possible that in the future we will have some none certified
connectors using the cdk so the change in this PR still makes sense.

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [x] YES 💚
- [ ] NO ❌